Create searching algorithm + relevant test cases
I wrote code for a very basic search algorithm that looks for
a search term in Supplier names. No docstring, no other features.

sample/tests directory serves as the home of any new test code.

Lastly, an addition to the dictionary so that PyCharm doesn't pick
"Yug" as a spelling error.