Transfer codebase to Django framework
We'll be using Django as our backend. I'm beginning to restructure
our code so that it's compatible with Django. This means I'm
temporarily removing the classes and test cases we've written. As
is, they will not work with Django and we need to rewrite them.

Backups of the repository have been created.