Convert product form view to use a form instead of a formset
This streamlines the efficiency of our code and fixes a bug that
would not show any forms at all if the database was empty.

I also wrote a recursive function that removes extra spaces
from the name field. I realized that only checking for double
spaces would not handle spaces of length 3+.